requirements.txt
torch>=2.2.0
torchvision>=0.17.0
open_clip_torch>=2.24.0
pillow>=10.0.0
pandas>=2.2.0
numpy>=1.24.0
tqdm>=4.66.0
faiss-cpu>=1.7.4
